Changed column price

+++ api/models.py
@@ -4,7 +4,7 @@ class Task(models.Model):
     id = models.AutoField(primary_key=True)
     title = models.CharField(max_length=200)
     description = models.TextField()
-    price = models.IntegerField(null=True)
+    price = models.FloatField(null=True)
     completed = models.BooleanField(default=False)
     created = models.DateTimeField(auto_now_add=True)
